refactor(error): 에러 처리 관련 클래스를 domain 패키지 밖으로 이동
