+++ src/main/java/roomescape/domain/reservation/entity/Reservation.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Objects;
-import roomescape.domain.global.exception.custom.BusinessException;
-import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.global.error.exception.BusinessException;
+import roomescape.global.error.ErrorCode;
 import roomescape.domain.theme.entity.Theme;
 import roomescape.domain.time.entity.Time;
