+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.global.exception.custom.BusinessException;
-import roomescape.domain.global.exception.error.ErrorCode;
-import roomescape.domain.global.exception.error.ErrorDetail;
+import roomescape.global.error.exception.BusinessException;
+import roomescape.global.error.ErrorCode;
+import roomescape.global.error.ErrorDetail;
 import roomescape.domain.reservation.dto.request.ReservationCreateRequestDto;
 import roomescape.domain.reservation.dto.request.ReservationUpdateRequestDto;
 import roomescape.domain.reservation.dto.response.ReservationCreateResponseDto;
