+++ src/main/java/roomescape/domain/theme/service/ThemeService.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.global.exception.custom.BusinessException;
-import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.global.error.exception.BusinessException;
+import roomescape.global.error.ErrorCode;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.theme.dto.request.ThemeCreateRequestDto;
 import roomescape.domain.theme.dto.response.ThemeResponseDto;
