+++ src/main/java/roomescape/domain/time/service/TimeService.java
@@ -5,9 +5,9 @@ import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.global.exception.custom.BusinessException;
-import roomescape.domain.global.exception.error.ErrorCode;
-import roomescape.domain.global.exception.error.ErrorDetail;
+import roomescape.global.error.exception.BusinessException;
+import roomescape.global.error.ErrorCode;
+import roomescape.global.error.ErrorDetail;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.theme.repository.ThemeRepository;
 import roomescape.domain.time.dto.request.TimeCreateRequestDto;
