+++ src/test/java/roomescape/domain/reservation/validator/ReservationValidatorTest.java
@@ -9,8 +9,8 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.custom.BusinessException;
-import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.global.error.exception.BusinessException;
+import roomescape.global.error.ErrorCode;
 import roomescape.domain.reservation.entity.Reservation;
 import roomescape.domain.theme.entity.Theme;
 import roomescape.domain.time.entity.Time;
